Add Rollback after failed TLS cert rotation when DB has not been updated (#1345)
In https://github.com/vertica/vertica-kubernetes/pull/1335, there are
three use cases for rollback from TLS cert rotation failure; this PR
covers case 1:

> https/clientserver cert rotation fails but before any change was made
in the tls config. We are still using the old cert so we just need to
undo the secret name change. (no-op on nma)

In this case, the DB has not been updated, so no second rotation is
required. We only require that the user revert their change to the spec,
in which case we can reset the conditions. Then we will be back in a
working state. So we will produce an Event alerting the user to revert
their change in order to trigger rollback. When the change has been
reverted, we will trigger rollback (which will just unset the conditions
in this case).

One additional change required for this is to change how/when we write
the secrets to the NMA TLS ConfigMap. Previously, this was done early,
which would require us to revert the change on rollback. However, now,
we will only update the ConfigMap when we are sure that cert rotation
has completed successfully.