Check for cert not matching key and for cert not matching common name (#1360)
Added two new cases to the function that will validate TLS certificates
for httpsNMA and clientServer secrets:

1. Validate if private key and certificate are matching; if not, produce
event and requeue.
2. Validate if subject.commonName of cert matches the database user; if
not, produce event and requeue. By default, the dbuser will be assumed
to be the annotation value vertica.com/superuser-name (default is
dbadmin). However, for clientServer, this can be overridden using the
new spec field below.

For httpsNMA, we will only ever validate against superuser-name.
However, for clientServer, we have added a new field to the spec section
clientServerTLS: commonName. When specified, clientServer secret's
commonName will be validated against this value instead of
superuser-name.