Automatically rotate through a list of TLS secrets using a set interval (#1363)
Prototype of automatic cert rotation. Includes a new section called
autoRotate in the TLSConfig sections of the spec (so, httpsNMATLS and
clientServerTLS). For example:

```
spec:
  clientServerTLS:
    autoRotate:
      interval: 30
      restartAtEnd: false
      secrets:
      - cert1
      - cert2
      - cert3
    mode: TRY_VERIFY
    secret: cert1
```

- secrets: the list of secrets to rotate through
- interval: how many days between rotations (default 30)
- restartAtEnd: when list is done, whether or not to start over again

This is done via a new reconciler at the end of our list, which will
requeue until the next auto-rotation is due. It also uses 2 new fields
in the TLSConfigs section of the status:

```
 status:
  tlsConfigs:
  - lastUpdate: "2025-07-31T16:00:06Z"
    mode: try_verify
    name: clientServer
    secret: cert1
    autoRotateSecrets:
    - cert1
    - cert2
    - cert3 
```

- autoRotateSecrets: copy of the secret list from the spec (in case spec
changes)
- lastUpdate: last time auto-rotate ran for this config; used to compute
next rotation

I also added an annotation for testing, since we don't want to wait a
day.

NOTE: does NOT include rollback cases

Still to do:

1. ~~Handle changes to secret list or interval~~ DONE: If current secret
is in new spec list, continue rotation with the new list; if not,
restart from scratch with first secret
2. ~~Webhooks for destructive cases, like empty secrets list or trying
to manually set secret when auto-rotate is scheduled~~ DONE: added
webhooks for empty or 1 element list, 0 interval, duplicate values.
However, manually changing secret is difficult; we can't prevent this
because auto-rotate will do exactly this. Future work?
3. Unit tests; code cleanup
4. If user sets autoRotate: in their spec and later removes it, the
auto-rotation will stop. However, the lastUpdate and autoRotateSecrets
will remain set in their status, which might be confusing. Remove it?

---------

Co-authored-by: Cai Chen <cchen8@opentext.com>