Fix server upgrade (#1373)
offline-upgrade-rename-cm has a timing issue. We used to delete the
current configmap (mounted), update the image first then replace the
deleted configmap with a new one. The issue is if offline upgrade starts
before the operator detects the cm change, the sts upgrade stratety will
be changed to OnDelete, leading to the pod being unable to update the
old configmap name (now deleted).
This updates the image and the cm in one call getting rid of that timing
issue. The goal of the test is to make sure a server upgrade,
accompanied with other CR changes, works. This change does not change
that.