Change rollback annotation from "disable" to "enable"; default to enabled (#1390)
For rollback after TLS cert rotation failure, previously, we had this
annotation:

`vertica.com/disable-tls-rotation-failure-rollback
`

This was defaulted to TRUE, disabling the feature. However, if a
customer wanted to enable this feature, they would have to change this
annotation to FALSE. This is awkward; to enable a new feature, they must
set a flag to false.

This PR will change the annotation to:

`vertica.com/enable-tls-rotation-failure-rollback
`

The default will still be TRUE, which means that this feature will now
be enabled by default. If a customer wants to disable it, they will set
the enable flag to FALSE, which makes more logical sense.