Support rollback after HTTPS cert rotation when TLS has been updated (#1386)
This is a clean branch and a simplification of this PR:
https://github.com/vertica/vertica-kubernetes/pull/1351

Currently, we are able to rollback after a failed Client-Server cert
rotation. For HTTPS, if the failure occurs **BEFORE** the TLS update
takes place in the DB, this is also supported. This PR will support
HTTPS rollback **AFTER** the DB has been updated; the most common cause
of such a rollback is a failure in the HTTPS polling step of cert
rotation.

In this rollback, we need to re-run the HTTPS cert rotation. However,
rather than rotating from status (previous values) to spec (new one), we
are rotating from spec (the failed secret) to status (the last good
secret). However, while the HTTPS cert has been rotated to the failed
one, the NMA has not yet. So, during the rollback rotation, we need to
point NMA to the status value (the last one).