Rollback after invalid cert error (#1397)
When you change the clientServer or httpsNMA secret, we will run several
checks to validate the secret's certificate. Previous, this failure
would not trigger a rollback; it would just produce errors until the
user changes to a valid secret. This is different than other cert
rotation failures, where we would rollback.

Now, when cert validation fails, we will trigger a rollback to the last
good secret. This will only apply when changing a secret; during create
db, it will still produce an error and expect the user to fix it (since
there is no last good secret).

ALSO: added a small bug fix for auto cert rotation. If you set auto cert
rotation then unset it, we would not clear out these fields in the
status. This means that, if you re-set auto cert rotation in the future,
this could lead to issues.