Save config object names to status (#1415)
This follows the work done in #1370. There was a bug. The list of
configmaps/secrets used to inject env vars were saved as a comma
separate list in vdb labels. That is wrong because first comma is not a
valid char for a label, and second labels are not used to keep this kind
of metadata.
Now I am following the right approach by saving the list in the status.