Add retry logic when HTTPS cert rotation fails during HTTPS polling (#1395)
At the end of the HTTPS cert rotation, after the DB has been updated, we
run HTTPS polling to verify that the HTTPS service is back up and
running with the new cert. If this fails, we will be forced to do a
rollback. However, HTTPS polling could potentially time out simply
because the HTTPS is slow to come up or due to some other transient
issue. Rather than rolling back in these situations, we want to give the
user the option to retry the HTTPS polling.

This adds a new annotation::
```
vertica.com/cert-rotation-max-retries:
```

This will allow the user is set how many retries to attempt, before
giving up and running rollback. This retry is only triggered if there is
an error specific to HTTPS polling; all other errors will rollback
immediately. The current retry is tracked in the status field
```httpsPollingCurrentRetry```.

To test this, I also updated the following testing annotation:

```
vertica.com/trigger-tls-update-failure:
```

To support a comma-separated list. This will allow us to trigger a
sequence of failures, in order to fully test retry.

---------

Co-authored-by: Libo Yu <libo.yu@microfocus.com>
Co-authored-by: LiboYu2 <168450736+LiboYu2@users.noreply.github.com>