Add ability to enable/disable TLS Auth by config (#1400)
Currently, TLS auth is enabled/disabled by an annotation
```enable-tls-auth```, which is set at the DB level. We want the ability
to enable or disable TLS auth individually for each TLS config (httpsNMA
and clientServer), instead of at the DB level.

We will add a new field ```enabled``` to the TLS Config sections of the
spec. For example:
```
spec:
  httpsNMATLS:
    secret: cert1
    enabled: true
```

It can be set to true or false, enabling and or disabling TLS auth for
this config. This also means de-coupling the logic within the operator,
so that we are checking each auth separately. For now, this is only
relevant when creating a new VDB or when first setting TLS config; a
webhook will prevent user from disabling a TLS config, once it has been
enabled.

Default for "enabled" will typically be True. However, the specific
functionality is:

1. If a TLS Config is defined in the spec and "enabled" is set to True
or False, we will honor that value (ignoring the annotation)
2. If a TLS Config is defined but "enabled" is not set, default will be
True; this is for backwards compat and to support the upgrade case
3. If a TLS Config is not defined, we will default to the annotation
value; in the future, we will remove the annotation and this case will
be False

One thing to consider is automatic cert generation. Currently
functionality is:

1. If TLS annotation is set to True and httpsNMATLS.secret or
clientServerTLS.secret are empty, we will auto-generate a secret for the
empty secret(s) and set the spec accordingly
2. If TLS annotation is false/unset and nmaTLSSecret is empty, we will
auto-generate a secret for this and set the spec accordingly
3. If TLS annotation is changed from False to True and nmaTLSSecret is
not empty, its value will be copied into httpsNMATLS.secret and/or
clientServerTLS.secret, if either of them are empty

New functionality is:

1. If httpsNMATLS.enabled is set to True and httpsNMATLS.secret is empty
and nmaTLSSecret is empty, we will auto-generate a secret for
httpsNMATLS.secret and set the spec accordingly. Similar for
clientServerTLS.
2. If httpsNMATLS.enabled is set to True and httpsNMATLS.secret is empty
and nmaTLSSecret is not empty, we will copy the value from nmaTLSSecret
into httpsNMATLS.secret. Similar for clientServerTLS.
3. If httpsNMATLS.enabled is set to False and nmaTLSSecret is empty, we
will auto-generate a secret for nmaTLSSEcret and set the spec
accordingly. This is unrelated to clientServerTLS.