Fix issues related cert gen rollback (#1430)
Improvements to make tlsservercertgen_reconciler more deterministic and
to avoid issues:

1. Previously, to generate/validate the 3 secrets, we were looping
through a Map, relying on ordering to make sure we generate the secrets
in the right order. However, apparently, iterating over Maps in Go is
not deterministic and ordering is not guaranteed. This would result in
things like generating the NMA secret first then pushing it to
clientServerTLS, rather than giving them two separate secrets. I have
changed this to a Struct, which apparently is deterministic.
2. We could get into a case where we trigger the rollback code, even
though there is no existing secret in the status. I added some more
protections against this.
3. In invalid-cert-validation, for clientServerTLS, we were setting
commonName but using an auto-generated secret (which did not use this
CN). This resulted in a validation error, which would sometimes mask the
actual validation error we were asserting in httpsNMATLS. So I gave
clientServerTLS a valid secret, matching the CN.
4. We had issues where a secret would fail validation but then we would
try to rotate to this secret anyway. This is a result of the TLS update
reconciler having a stale VDB. I believe this could be a symptom of 1
above. I can't reproduce this anymore, but I have added some additional
logging, in case we hit this in the future.