Change password in sandbox and other places when passwordSecret is updated (#1410)
Previously, if passwordSecret in spec was updated and VDB contained
sandboxes, this was prevented via webhook. Now, after updating the main
cluster, we will loop through each sandbox and update their passwords as
well. However, if any sandbox is shutdown and passwordSecret is changed,
we will still block this via webhook (since we cannot update them).

Also, we want any reconciler that needs to get DB password to share
password logic and password cache. This will add a PasswordManager,
which provides a shared password cache across all reconcilers that need
it.