Set tls enabled after operator upgrade (#1431)
Previously If you were in say 25.3.0 with enable tls annotation set to
"false" and you upgrade to 25.4.0, because the tls struct is set, we
would enable (https/client-server) configs.
Now, we detect if we just did an upgrade by comparing the operator
version in statefulsets to the current version, if they do not match, we
use the enable tls annotation to set the tls config spec. If the tls
config spec is nil or if enabled is not nil then there is nothing to do.