Fix issues related to password change in sandbox (#1437)
ISSUE 1

We have a webhook that prevents sandboxing/unsandboxing when a password
change is in progress. This webhook is incorrectly blocking creation of
VDBs with both sandboxes and passwordSecret set. The reason this is
happening is because passwordSecret in spec technically does not match
passwordStatus in status (which is nil).

The fix is to not trigger this webhook when passwordStatus is nil. The
only time passwordSecret in status will be nil is before
passwordsecret_reconciler has first run, when it will initialize
status.passwordSecret, changing it from nil->"" or
nil->"some_passsword".

ISSUE 2

Creating VDB with no password and then setting password could cause
operator to get stuck, using wrong password. This is related to the
initialization logic. The empty passwordSecret would never get
initialized in the status. Therefore, when the password was changed, it
assumed this was initialization; it set the status but did not update
the DB.

To fix this, I have separated "password has been changed" logic from
"password has not been initialized" logic.

ISSUE 3

These webhooks would be skipped when "sandboxes:" in spec was empty,
since they are all related to sandboxes. However, this created an issue
when unsandboxing your last sandbox. They will now skip when
"sandboxes:" is empty for BOTH VDBs: before and after.