Add fix for upgrade issue (#1441)
In older operator versions like v25.3.0, spec.nmaTLSSecret was
deprecated. We would always set spec.httpsNMATLS.secret even when tls is
disabled. If you upgrade to 25.4.0 or later with tls disabled,
spec.nmaTLSSecret is back and is autogerated if empty.
The issue is when tls is disabled, nma in v25.3.0 got its secret from
spec.httpsNMATLS.secret while in v25.4.0+, it assumes the secret is in
spec.nmaTLSSecret.
To honore that, instead of generating a new secret, we will copy
spec.httpsNMATLS.secret to spec.nmaTLSSecret.