Fail Community Edition license deployment (#1411)
Here is the design doc for this change:

https://confluence.verticacorp.com/display/DEV/Invalidate+Vertica+Community+Edition+%28CE%29+License+in+Kubernetes+Environment

Here is the related PR in gitlab repo: 

https://gitlab.otxlab.net/analytics-database/ver/server/-/merge_requests/23594

This is what we want to achieve:
1 If users want to create a new database using CE license, the operator
will reject the license.
2 For existing databases created without a license secret or with a CE
license, the operator will stop to reconcile.

How to validate a license?
run vertica binary with "-L" flag. For CE license, its outlook looks
like this:
That Vertica license is valid
Vertica Community Edition 
2011-11-22
Perpetual
1TB CE Nodes 3
Total Entitlements:
Capacity: 1TB
Nodes: 3

We will use that "Vertica Community Edition " field to identify a CE
license and reject it.

NMA already has an endpoint that uses "-L" flag. A vclusterops API has
been added to allow the operator to validate a license secret using that
NMA endpoint.

Reconcile logic:
1 For failed license validation, an error is returned.
2 For successful license validation, the timestamp is saved into Status.
And the reconciler will validate the license
on a daily basis by default. A new annotation is added to control that
validation interval. The maximum interval is one day.

---------

Co-authored-by: Roy Paulin <rnguetsopken@opentext.com>